+++ src/main/java/me/anedhel/lotr/block/ModStoneBlocks.java
@@ -1999,6 +1999,200 @@ public class ModStoneBlocks {
 			.wall(CRACKED_OVERGROWN_GOLD_FANCY_BLUESLATE_BRICKS_WALL)
 			.group("stone").unlockCriterionName("has_cracked_gold_fancy_bricks").build();
 
+	// CHALK BLOCKS - Following the same pattern as blueslate
+	public static final Block CHALK = registerBlock("chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block CHALK_STAIRS = registerBlock("chalk_stairs",
+			new StairsBlock(CHALK.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block CHALK_SLAB = registerBlock("chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block CHALK_WALL = registerBlock("chalk_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+	public static final Block CHALK_BUTTON = registerBlock("chalk_button",
+			new ButtonBlock(FabricBlockSettings.copyOf(Blocks.STONE_BUTTON), BlockSetType.STONE, 10, false));
+	public static final Block CHALK_PRESSURE_PLATE = registerBlock("chalk_pressure_plate",
+			new PressurePlateBlock(PressurePlateBlock.ActivationRule.EVERYTHING,
+					FabricBlockSettings.copyOf(Blocks.STONE_PRESSURE_PLATE), BlockSetType.STONE));
+	public static final BlockFamily CHALK_FAMILY = BlockFamilies.register(CHALK)
+			.stairs(CHALK_STAIRS)
+			.slab(CHALK_SLAB)
+			.wall(CHALK_WALL)
+			.button(CHALK_BUTTON)
+			.pressurePlate(CHALK_PRESSURE_PLATE)
+			.group("stone").unlockCriterionName("has_stone").build();
+
+	public static final Block COBBLED_CHALK = registerBlock("cobbled_chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block COBBLED_CHALK_STAIRS = registerBlock("cobbled_chalk_stairs",
+			new StairsBlock(COBBLED_CHALK.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block COBBLED_CHALK_SLAB = registerBlock("cobbled_chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block COBBLED_CHALK_WALL = registerBlock("cobbled_chalk_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+	public static final Block COBBLED_CHALK_BUTTON = registerBlock("cobbled_chalk_button",
+			new ButtonBlock(FabricBlockSettings.copyOf(Blocks.STONE_BUTTON), BlockSetType.STONE, 10, false));
+	public static final Block COBBLED_CHALK_PRESSURE_PLATE = registerBlock("cobbled_chalk_pressure_plate",
+			new PressurePlateBlock(PressurePlateBlock.ActivationRule.EVERYTHING,
+					FabricBlockSettings.copyOf(Blocks.STONE_PRESSURE_PLATE), BlockSetType.STONE));
+	public static final BlockFamily COBBLED_CHALK_FAMILY = BlockFamilies.register(COBBLED_CHALK)
+			.stairs(COBBLED_CHALK_STAIRS)
+			.slab(COBBLED_CHALK_SLAB)
+			.wall(COBBLED_CHALK_WALL)
+			.button(COBBLED_CHALK_BUTTON)
+			.pressurePlate(COBBLED_CHALK_PRESSURE_PLATE)
+			.group("stone").unlockCriterionName("has_cobbled_stone").build();
+
+	public static final Block COBBLED_CHALK_BRICKS = registerBlock("cobbled_chalk_bricks",
+			new Block(FabricBlockSettings.copyOf(Blocks.COBBLESTONE)));
+	public static final Block COBBLED_CHALK_BRICKS_STAIRS = registerBlock("cobbled_chalk_bricks_stairs",
+			new StairsBlock(COBBLED_CHALK_BRICKS.getDefaultState(), FabricBlockSettings.copyOf(Blocks.COBBLESTONE_STAIRS)));
+	public static final Block COBBLED_CHALK_BRICKS_SLAB = registerBlock("cobbled_chalk_bricks_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.COBBLESTONE_SLAB)));
+	public static final Block COBBLED_CHALK_BRICKS_WALL = registerBlock("cobbled_chalk_bricks_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.COBBLESTONE_WALL)));
+	public static final BlockFamily COBBLED_CHALK_BRICKS_FAMILY = BlockFamilies.register(COBBLED_CHALK_BRICKS)
+			.stairs(COBBLED_CHALK_BRICKS_STAIRS)
+			.slab(COBBLED_CHALK_BRICKS_SLAB)
+			.wall(COBBLED_CHALK_BRICKS_WALL)
+			.group("stone").unlockCriterionName("has_cobbled_brick").build();
+	public static final Block CRACKED_COBBLED_CHALK_BRICKS = registerBlock("cracked_cobbled_chalk_bricks",
+			new Block(FabricBlockSettings.copyOf(Blocks.COBBLESTONE)));
+	public static final Block CRACKED_COBBLED_CHALK_BRICKS_STAIRS = registerBlock("cracked_cobbled_chalk_bricks_stairs",
+			new StairsBlock(CRACKED_COBBLED_CHALK_BRICKS.getDefaultState(), FabricBlockSettings.copyOf(Blocks.COBBLESTONE_STAIRS)));
+	public static final Block CRACKED_COBBLED_CHALK_BRICKS_SLAB = registerBlock("cracked_cobbled_chalk_bricks_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.COBBLESTONE_SLAB)));
+	public static final Block CRACKED_COBBLED_CHALK_BRICKS_WALL = registerBlock("cracked_cobbled_chalk_bricks_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.COBBLESTONE_WALL)));
+	public static final BlockFamily CRACKED_COBBLED_CHALK_BRICKS_FAMILY = BlockFamilies.register(CRACKED_COBBLED_CHALK_BRICKS)
+			.stairs(CRACKED_COBBLED_CHALK_BRICKS_STAIRS)
+			.slab(CRACKED_COBBLED_CHALK_BRICKS_SLAB)
+			.wall(CRACKED_COBBLED_CHALK_BRICKS_WALL)
+			.group("stone").unlockCriterionName("has_cracked_cobbled_brick").build();
+
+	public static final Block SMOOTH_CHALK = registerBlock("smooth_chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE)));
+	public static final Block SMOOTH_CHALK_SLAB = registerBlock("smooth_chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE_SLAB)));
+	public static final Block CRACKED_SMOOTH_CHALK = registerBlock("cracked_smooth_chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE)));
+	public static final Block CRACKED_SMOOTH_CHALK_SLAB = registerBlock("cracked_smooth_chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE_SLAB)));
+
+	public static final Block CHALK_BRICK = registerBlock("chalk_brick",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block CHALK_BRICK_STAIRS = registerBlock("chalk_brick_stairs",
+			new StairsBlock(CHALK_BRICK.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block CHALK_BRICK_SLAB = registerBlock("chalk_brick_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block CHALK_BRICK_WALL = registerBlock("chalk_brick_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+	public static final Block CHISELED_CHALK_BRICK = registerBlock("chiseled_chalk_brick",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final BlockFamily CHALK_BRICK_FAMILY = BlockFamilies.register(CHALK_BRICK)
+			.stairs(CHALK_BRICK_STAIRS)
+			.slab(CHALK_BRICK_SLAB)
+			.wall(CHALK_BRICK_WALL)
+			.chiseled(CHISELED_CHALK_BRICK)
+			.group("stone").unlockCriterionName("has_brick").build();
+	public static final Block CRACKED_CHALK_BRICK = registerBlock("cracked_chalk_brick",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block CRACKED_CHALK_BRICK_STAIRS = registerBlock("cracked_chalk_brick_stairs",
+			new StairsBlock(CRACKED_CHALK_BRICK.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block CRACKED_CHALK_BRICK_SLAB = registerBlock("cracked_chalk_brick_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block CRACKED_CHALK_BRICK_WALL = registerBlock("cracked_chalk_brick_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+	public static final BlockFamily CRACKED_CHALK_BRICK_FAMILY = BlockFamilies.register(CRACKED_CHALK_BRICK)
+			.stairs(CRACKED_CHALK_BRICK_STAIRS)
+			.slab(CRACKED_CHALK_BRICK_SLAB)
+			.wall(CRACKED_CHALK_BRICK_WALL)
+			.group("stone").unlockCriterionName("has_cracked_brick").build();
+
+	public static final Block CHALK_TILES = registerBlock("chalk_tiles",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block CHALK_TILES_STAIRS = registerBlock("chalk_tiles_stairs",
+			new StairsBlock(CHALK_TILES.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block CHALK_TILES_SLAB = registerBlock("chalk_tiles_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block CHALK_TILES_WALL = registerBlock("chalk_tiles_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+	public static final BlockFamily CHALK_TILES_FAMILY = BlockFamilies.register(CHALK_TILES)
+			.stairs(CHALK_TILES_STAIRS)
+			.slab(CHALK_TILES_SLAB)
+			.wall(CHALK_TILES_WALL)
+			.group("stone").unlockCriterionName("has_tile").build();
+	public static final Block CRACKED_CHALK_TILES = registerBlock("cracked_chalk_tiles",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block CRACKED_CHALK_TILES_STAIRS = registerBlock("cracked_chalk_tiles_stairs",
+			new StairsBlock(CRACKED_CHALK_TILES.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block CRACKED_CHALK_TILES_SLAB = registerBlock("cracked_chalk_tiles_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block CRACKED_CHALK_TILES_WALL = registerBlock("cracked_chalk_tiles_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+	public static final BlockFamily CRACKED_CHALK_TILES_FAMILY = BlockFamilies.register(CRACKED_CHALK_TILES)
+			.stairs(CRACKED_CHALK_TILES_STAIRS)
+			.slab(CRACKED_CHALK_TILES_SLAB)
+			.wall(CRACKED_CHALK_TILES_WALL)
+			.group("stone").unlockCriterionName("has_cracked_tile").build();
+
+	public static final Block CHALK_PILLAR = registerBlock("chalk_pillar",
+			new ModPillarBlock(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block CHALK_PILLAR_SLAB = registerBlock("chalk_pillar_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block CRACKED_CHALK_PILLAR = registerBlock("cracked_chalk_pillar",
+			new ModPillarBlock(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block CRACKED_CHALK_PILLAR_SLAB = registerBlock("cracked_chalk_pillar_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+
+	public static final Block CHALK_PAVEMENT = registerBlock("chalk_pavement",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block CHALK_PAVEMENT_STAIRS = registerBlock("chalk_pavement_stairs",
+			new StairsBlock(CHALK_PAVEMENT.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block CHALK_PAVEMENT_SLAB = registerBlock("chalk_pavement_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final BlockFamily CHALK_PAVEMENT_FAMILY = BlockFamilies.register(CHALK_PAVEMENT)
+			.stairs(CHALK_PAVEMENT_STAIRS)
+			.slab(CHALK_PAVEMENT_SLAB)
+			.group("stone").unlockCriterionName("has_pavement").build();
+	public static final Block CRACKED_CHALK_PAVEMENT = registerBlock("cracked_chalk_pavement",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block CRACKED_CHALK_PAVEMENT_STAIRS = registerBlock("cracked_chalk_pavement_stairs",
+			new StairsBlock(CRACKED_CHALK_PAVEMENT.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block CRACKED_CHALK_PAVEMENT_SLAB = registerBlock("cracked_chalk_pavement_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final BlockFamily CRACKED_CHALK_PAVEMENT_FAMILY = BlockFamilies.register(CRACKED_CHALK_PAVEMENT)
+			.stairs(CRACKED_CHALK_PAVEMENT_STAIRS)
+			.slab(CRACKED_CHALK_PAVEMENT_SLAB)
+			.group("stone").unlockCriterionName("has_cracked_pavement").build();
+
+	public static final Block FANCY_CHALK_BRICKS = registerBlock("fancy_chalk_bricks",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block FANCY_CHALK_BRICKS_STAIRS = registerBlock("fancy_chalk_bricks_stairs",
+			new StairsBlock(FANCY_CHALK_BRICKS.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block FANCY_CHALK_BRICKS_SLAB = registerBlock("fancy_chalk_bricks_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block FANCY_CHALK_BRICKS_WALL = registerBlock("fancy_chalk_bricks_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+	public static final BlockFamily FANCY_CHALK_BRICKS_FAMILY = BlockFamilies.register(FANCY_CHALK_BRICKS)
+			.stairs(FANCY_CHALK_BRICKS_STAIRS)
+			.slab(FANCY_CHALK_BRICKS_SLAB)
+			.wall(FANCY_CHALK_BRICKS_WALL)
+			.group("stone").unlockCriterionName("has_fancy_bricks").build();
+	public static final Block CRACKED_FANCY_CHALK_BRICKS = registerBlock("cracked_fancy_chalk_bricks",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block CRACKED_FANCY_CHALK_BRICKS_STAIRS = registerBlock(
+			"cracked_fancy_chalk_bricks_stairs",
+			new StairsBlock(CRACKED_FANCY_CHALK_BRICKS.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block CRACKED_FANCY_CHALK_BRICKS_SLAB = registerBlock("cracked_fancy_chalk_bricks_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block CRACKED_FANCY_CHALK_BRICKS_WALL = registerBlock("cracked_fancy_chalk_bricks_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+	public static final BlockFamily CRACKED_FANCY_CHALK_BRICKS_FAMILY = BlockFamilies.register(CRACKED_FANCY_CHALK_BRICKS)
+			.stairs(CRACKED_FANCY_CHALK_BRICKS_STAIRS)
+			.slab(CRACKED_FANCY_CHALK_BRICKS_SLAB)
+			.wall(CRACKED_FANCY_CHALK_BRICKS_WALL)
+			.group("stone").unlockCriterionName("has_cracked_fancy_bricks").build();
+
     /**
      * This Method is used to register a Block with minecraft and also register the Item for the block
      * @param name the name of the Block